ConvertModels/requirements.txt
coremltools==7.1
transformers==4.36.0
torch==2.1.0
sentencepiece==0.1.99
pyyaml==6.0.1
einops==0.7.0
numpy<2.0.0
